shill: 64-bit fixes: integer sizes for printf
BUG=chromium-os:23594
TEST=Rerun unit tests (on x86 and amd64)

Reviewed-on: https://gerrit.chromium.org/gerrit/12283
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>

+++ shill/shill_unittest.cc
@@ -76,8 +76,9 @@ class MockEventDispatchTester {
 
   void HandleData(InputData *inputData) {
     LOG(INFO) << "MockEventDispatchTester handling data len "
-              << base::StringPrintf("%d %.*s", inputData->len,
-                                    inputData->len, inputData->buf);
+              << base::StringPrintf("%zd %.*s", inputData->len,
+                                    static_cast<int>(inputData->len),
+                                    inputData->buf);
     got_data_ = true;
     IOComplete(inputData->len);
     StopDispatcher();
